feat(comms): fixed feedback hysteresis, added extraction capabilities for four arguments
The feedback calculations had a mistake in them, causing the motor to
output little to no speed when the error was in the clockwise
direction. Also added the capability for the RPi to unpack four values
(encoder count, rotation direction, current, and homing boolean).

Written by: V.S. Boon
Committed on: 15-10-2021
Version: 0.3.6