feat(teensy): Rewrite serial comm for Teensy
The new serial_comm.ino is tailored towards the expected pinout of the Teensy 4.0, as well as utilizing teniary functions and arrays in order to obtain a cleaner and less bloated code.
For even briefer code, it might be worth looking into lambda functions to allow for the interrupt functions to 'take inputs'.

Written by: V.S. Boon
Committed on: 03-11-2021
Version: 0.4.2