feat(enc): low-pass filtering
The encoder reader now has a low-pass filter with a cut-off frequency of 1/errTime, typically in the order of 1 MHz, with errTime being in the order of a microsecond. This is to avoid interference being read as encoder counts. Moreover, call code for current sensing is commented out, as the first prototype system does not support it.

Written by: V.S. Boon
Committed on: 16-11-2021
Version: 0.5.2