fix: manual control
Fix includes, but is not limited to, fixing datatypes for feed-forward & PID and fixing parsing issues caused by incorrect set dataPacket length.
KNOWN CURRENT ISSUE: When a button is pressed, it is not tracked if this button stays pressed, which causes the code to start running HoldPos instantly after pushing a button. A solution could be a dictionary with booleans as values, which get flipped at a key press and -release.

Written by: V.S. boon
Committed on: 17-11-2021
Version: 0.5.4