fix: manual control & serial bugs
Main problems had to do with the serial data being six 'packets' (one for each motor-encoder pair), while the current system worked with a serial data length of 5. This length was consequently used to do the many serial robotics-based calculations, which require the number of joints (5). Hence, the code had to be slightly refactored and rewritten.

Additionally, the speed-control bug, where a button press would only result in an increase in velocity for its particular frame, has been fixed. Some additional fixes were made to the end-effector control, although this is not yet fully solved.

Written by: V.S. Boon
Committed on: 17-11-2021
Version: 0.5.5