fix(comm): direction parsing & mSpeed changing
The robot is now operable in the joint-space. However, a known issue is that the current motor speeds are, as is to-be expected, for each motor. In the next update, the motor speeds (and joint limits etc.) should be in the joint space, i.e. the differential drive motors should work together.

Written by: V.S. Boon
Committed on: 18-11-2021
Version: 0.5.6