feat(main): Main control function
The main control function is the last piece of the puzzle for this first iteration of the PegasusArm OS. It combines all control methods in one handy function, which can be continuously (re-)run.

As of this update, the position control has been bug-fixed, atleast as far as unit tests etc. go: The robot has not been connected to this system, so the outputs might be significantly off.

Written by: V.S. Boon
Committed on: 12-12-2021
Version: 0.7.0