fix(velCtrl): velocity control in main
The CheckKeys() functions had to be rewritten, as detecting key presses / releases was too unrealiable. Instead, pygame.key.get_pressed() is used, which returns a list of booleans for each key to see if they are pressed or not, which is the same functionality I tried to achieve with my 'pressed' dictionary. Other bugs were also fixed.

Written by: V.S. Boon
Comitted on: 13-12-2021
Version: 0.7.1