fix: Conversion factors based on experiments.
Be cautious with using these values, as they have only been tested on one joint.

More reliable might be to directly use the conversion factor tau to PWM of
33.78 [1/(N/m)].

Take into account, however, that the theoretical conversion factors are in the order of 100 to 1.000 lower!

Written by: V.S. Boon
Committed on: 18-12-2021
Version: 0.7.3