perf: special dynamics joint 2, removed friction
During testing it was observed that the second joint seemingly overcompensated for the effect of the joint accelerations / forces of the joints above and below it. Hence, these effects have been nullified, but can be reinstated with ease.

Moreover, the calculation of friction has been moved outside of the feed-forward. This is such that if the PID wants to move the joint against the direction of the Feed-Forward, friction can still be used to obtain this.

Written by: V.S. Boon
Committed on: 18-12-2021
Version: 0.7.3